refactor: create overlay form util
